+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.10
 - Added flashback button to main menu
+- Added modmenu button to main menu
 - Added realms to main menu
 - Now shows e4mc hint as hover tooltip
 - Corrected punctuation in Mojang copyright